feat: working on merging the modal
