--- packages/api-client-modal/CHANGELOG.md
@@ -1,152 +1,0 @@
-# @scalar/api-client-modal
-
-## 0.0.16
-
-### Patch Changes
-
-- 94e68ab: chore: upgrade typescript to 5.5
-- Updated dependencies [3ab472b]
-- Updated dependencies [50c2944]
-- Updated dependencies [a298195]
-- Updated dependencies [326bd3d]
-- Updated dependencies [4aa74e3]
-- Updated dependencies [a298195]
-- Updated dependencies [94e68ab]
-- Updated dependencies [5103023]
-- Updated dependencies [1734dd6]
-  - @scalar/client-app@0.1.14
-  - @scalar/oas-utils@0.2.5
-  - @scalar/components@0.12.4
-  - @scalar/object-utils@1.1.2
-
-## 0.0.15
-
-### Patch Changes
-
-- Updated dependencies [132acd4]
-- Updated dependencies [3b53db5]
-- Updated dependencies [b981770]
-- Updated dependencies [16e4794]
-- Updated dependencies [fd18932]
-- Updated dependencies [06471ed]
-  - @scalar/client-app@0.1.13
-  - @scalar/components@0.12.3
-
-## 0.0.14
-
-### Patch Changes
-
-- Updated dependencies [65f7bf5]
-- Updated dependencies [4643bfa]
-- Updated dependencies [ee1e97e]
-  - @scalar/components@0.12.2
-  - @scalar/client-app@0.1.12
-
-## 0.0.13
-
-### Patch Changes
-
-- Updated dependencies [c20c7d0]
-- Updated dependencies [c20c7d0]
-- Updated dependencies [b3ba2a5]
-- Updated dependencies [c20c7d0]
-  - @scalar/components@0.12.1
-  - @scalar/client-app@0.1.11
-  - @scalar/oas-utils@0.2.4
-
-## 0.0.12
-
-### Patch Changes
-
-- Updated dependencies [e4419ce]
-- Updated dependencies [5812c2f]
-- Updated dependencies [f2c1019]
-- Updated dependencies [fd0c93d]
-- Updated dependencies [7b87b8c]
-- Updated dependencies [512c815]
-  - @scalar/components@0.12.0
-  - @scalar/client-app@0.1.10
-
-## 0.0.11
-
-### Patch Changes
-
-- Updated dependencies [5e92eab]
-  - @scalar/client-app@0.1.9
-  - @scalar/components@0.11.6
-
-## 0.0.10
-
-### Patch Changes
-
-- Updated dependencies [362d47a]
-- Updated dependencies [627e0e5]
-  - @scalar/oas-utils@0.2.3
-  - @scalar/components@0.11.5
-  - @scalar/client-app@0.1.8
-
-## 0.0.9
-
-### Patch Changes
-
-- Updated dependencies [3d6ca80]
-- Updated dependencies [0fae180]
-- Updated dependencies [d982f18]
-  - @scalar/components@0.11.4
-  - @scalar/client-app@0.1.7
-
-## 0.0.8
-
-### Patch Changes
-
-- Updated dependencies [7dbb241]
-  - @scalar/components@0.11.3
-  - @scalar/client-app@0.1.6
-
-## 0.0.7
-
-### Patch Changes
-
-- a2cb3c3: Migrate away from unintended peer dependencies
-- Updated dependencies [a2cb3c3]
-  - @scalar/client-app@0.1.5
-  - @scalar/components@0.11.2
-  - @scalar/oas-utils@0.2.2
-
-## 0.0.6
-
-### Patch Changes
-
-- @scalar/client-app@0.1.4
-
-## 0.0.5
-
-### Patch Changes
-
-- a9b2395: chore: make api client modal public
-- e0fc110: chore: patch bump all package
-- Updated dependencies [be3a679]
-- Updated dependencies [e0fc110]
-- Updated dependencies [f0acc89]
-  - @scalar/client-app@0.1.3
-  - @scalar/components@0.11.1
-  - @scalar/oas-utils@0.2.1
-
-## 0.0.2
-
-### Patch Changes
-
-- 8494349: Migrate to highlightjs based syntax highlighting
-- Updated dependencies [190bcb6]
-- Updated dependencies [7f8ef74]
-- Updated dependencies [f82a767]
-- Updated dependencies [326dcfb]
-- Updated dependencies [96347a1]
-- Updated dependencies [8494349]
-- Updated dependencies [b38c7ed]
-- Updated dependencies [6fbb57e]
-- Updated dependencies [94084bf]
-- Updated dependencies [c5760b9]
-  - @scalar/components@0.11.0
-  - @scalar/client-app@0.1.0
-  - @scalar/oas-utils@0.2.0
