--- packages/api-client-modal/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@scalar/api-client-modal",
-  "description": "A modal wrapper around the API Client",
-  "license": "MIT",
-  "author": "Scalar (https://github.com/scalar)",
-  "homepage": "https://github.com/scalar/scalar",
-  "bugs": "https://github.com/scalar/scalar/issues/new/choose",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/scalar/scalar.git",
-    "directory": "packages/api-client-modal"
-  },
-  "keywords": [
-    "vue api client modal"
-  ],
-  "version": "0.0.16",
-  "engines": {
-    "node": ">=18"
-  },
-  "scripts": {
-    "build": "vite build && pnpm types:build && tsc-alias -p tsconfig.build.json",
-    "dev": "vite",
-    "lint:check": "eslint .",
-    "lint:fix": "eslint .  --fix",
-    "preview": "vite preview",
-    "test": "vitest",
-    "types:build": "vue-tsc -p tsconfig.build.json",
-    "types:check": "vue-tsc --noEmit --skipLibCheck --composite false"
-  },
-  "type": "module",
-  "main": "dist/index.js",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    },
-    "./components": {
-      "import": "./dist/components/index.js",
-      "types": "./dist/components/index.d.ts"
-    },
-    "./css/*.css": {
-      "import": "./dist/css/*.css",
-      "require": "./dist/css/*.css"
-    },
-    "./*.css": {
-      "import": "./dist/*.css",
-      "require": "./dist/*.css"
-    }
-  },
-  "files": [
-    "dist",
-    "CHANGELOG.md"
-  ],
-  "imports": {
-    "#@/*": "./src/*.ts"
-  },
-  "module": "dist/index.js",
-  "dependencies": {
-    "@scalar/client-app": "workspace:*",
-    "@scalar/components": "workspace:*",
-    "@scalar/oas-utils": "workspace:*",
-    "@scalar/object-utils": "workspace:*",
-    "vue": "^3.4.22",
-    "vue-router": "^4.3.0"
-  },
-  "devDependencies": {
-    "@scalar/build-tooling": "workspace:*",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "tsc-alias": "^1.8.8",
-    "type-fest": "^4.20.0",
-    "vite": "^5.2.10",
-    "vue-tsc": "^2.0.13"
-  }
-}
