--- packages/api-client-modal/src/index.ts
@@ -1,2 +1,0 @@
-export * from './api-client-modal'
-export * from './types'
