--- packages/api-client-modal/vite.config.ts
@@ -1,23 +1,0 @@
-import { createViteBuildOptions, findEntryPoints } from '@scalar/build-tooling'
-import vue from '@vitejs/plugin-vue'
-import { URL, fileURLToPath } from 'node:url'
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  plugins: [vue()],
-  resolve: {
-    alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url)),
-    },
-    dedupe: ['vue'],
-  },
-  server: {
-    port: 5066,
-  },
-  build: createViteBuildOptions({
-    entry: await findEntryPoints({ allowCss: true }),
-    options: {
-      ssr: false,
-    },
-  }),
-})
