+++ packages/client-app/playground/index.html
@@ -12,10 +12,10 @@
     <title>Scalar API Client Modal</title>
   </head>
 
-  <body class="light-mode">
+  <body class="light-mode scalar-app scalar-client">
     <div id="root"></div>
     <script
       type="module"
-      src="/playground/main.ts"></script>
+      src="./main.ts"></script>
   </body>
 </html>
