+++ packages/client-app/src/Modal/index.ts
@@ -1,1 +1,2 @@
 export { default as ApiClientModal } from './ApiClientModal.vue'
+export * from './api-client-modal'
