+++ packages/client-app/src/index.ts
@@ -1,4 +1,5 @@
 export { default as ScalarClient } from './App.vue'
+export * from './Modal'
 
 export { useWorkspace } from './store/workspace'
