+++ packages/client-app/tsconfig.json
@@ -6,5 +6,5 @@
       "@test/*": ["./test/*"]
     }
   },
-  "include": ["src", "global.d.ts"]
+  "include": ["src", "global.d.ts", "playground"]
 }
